refactor: simplify authRoutes and userRoutes by modularizing authController
